Removed openssl elements and fixed some bugs
Should be able to make this version.  Many of the input bugs are fixed
and some more clear messages have been added. passwords of length 24 are
causing passwords. Those of length 23 are not. Passwords of length 5 no
longer work and symbols can no longer be used as the last element of
names.